fix(model-selection): exclude models in quality escalation path and DRY provider tier lookup
The quality escalation path in Models::Select did not check
excluded_model_ids, allowing a project-excluded model to be selected
via quality recovery. Additionally, provider_tier_model and
excluded_model? were duplicated across three classes.

Extract shared ProviderTierLookup module and add the missing exclusion
check in quality_escalation_result.
